move volume image wrapper to gfxopengl
