+++ renderlib/gfxOpenGL/ImageXyzcGpu.cpp
